Seperate DefaultShardManager into Default and Abstract
